requirements.txt
goose3==3.1.6
hiredis==1.0.0
praw==6.1.1
pysummarize==0.6.0
redis==3.2.1
rfc3986==1.3.1